refactor: rename and replace main page components to widgets folder
