+++ src/pages/main/ui/MainPage.module.scss
