+++ src/widgets/main/ui/catalogMain/CatalogMain.module.scss
